Format & Clippy

+++ fast3d/src/gbi/utils.rs
@@ -1,3 +1,5 @@
+use crate::gbi::defines::RSP_GEOMETRY;
+use crate::rsp::RSPConstants;
 use crate::{
     models::texture::TextFilt,
     output::gfx::{BlendFactor, Face},
@@ -6,19 +8,17 @@ use crate::{
         OtherModeLayoutL,
     },
 };
-use crate::gbi::defines::RSP_GEOMETRY;
-use crate::rsp::RSPConstants;
 
 pub fn get_cmd(val: usize, start_bit: u32, num_bits: u32) -> usize {
     (val >> start_bit) & ((1 << num_bits) - 1)
 }
 
 pub fn geometry_mode_uses_lighting(geometry_mode: u32) -> bool {
-    geometry_mode & RSP_GEOMETRY::G_LIGHTING as u32 > 0
+    geometry_mode & RSP_GEOMETRY::G_LIGHTING > 0
 }
 
 pub fn geometry_mode_uses_fog(geometry_mode: u32) -> bool {
-    geometry_mode & RSP_GEOMETRY::G_FOG as u32 > 0
+    geometry_mode & RSP_GEOMETRY::G_FOG > 0
 }
 
 pub fn other_mode_l_uses_texture_edge(other_mode_l: u32) -> bool {
